# contracts/openapi/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "GruaHub API",
    "description" : "Plataforma multitenant para gestão de máquinas de pelúcia",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
...
  "paths" : {
...
    "/api/v1/visits" : {
      "get" : {
        "tags" : [ "Field Visit Resource" ],
        "parameters" : [ {
          "name" : "operatingPointId",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : 50,
            "type" : "integer"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseVisitResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FIELD_OPERATOR", "FINANCE" ]
        } ]
      },
      "post" : {
        "tags" : [ "Field Visit Resource" ],
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StartVisitRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FIELD_OPERATOR" ]
        } ]
      }
    },
    "/api/v1/visits/checklist" : {
      "post" : {
        "tags" : [ "Field Visit Resource" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChecklistRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FIELD_OPERATOR" ]
        } ]
      }
    },
    "/api/v1/visits/complete" : {
      "post" : {
        "tags" : [ "Field Visit Resource" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompleteVisitRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FIELD_OPERATOR" ]
        } ]
      }
    },
    "/api/v1/visits/{id}" : {
      "get" : {
        "tags" : [ "Field Visit Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VisitResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FIELD_OPERATOR", "FINANCE" ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ChecklistItem" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "checked" : {
            "type" : "boolean"
          }
        }
      },
      "ChecklistRequest" : {
        "required" : [ "visitId", "clientOperationId" ],
        "type" : "object",
        "properties" : {
          "visitId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "clientOperationId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ChecklistItem"
            }
          },
          "completedAt" : {
            "type" : "string"
          }
        }
      },
      "CompleteVisitRequest" : {
        "required" : [ "visitId", "clientOperationId" ],
        "type" : "object",
        "properties" : {
          "visitId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "clientOperationId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "checkoutAt" : {
            "type" : "string"
          },
          "cashCollectedCents" : {
            "format" : "int64",
            "type" : "integer"
          }
        }
      },
...
      "Instant" : {
        "format" : "date-time",
        "type" : "string",
        "example" : "2022-03-10T16:15:50Z"
      },
...
      "PageResponseVisitResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/VisitResponse"
            }
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          }
        }
      },
...
      "StartVisitRequest" : {
        "required" : [ "visitId", "clientOperationId", "operatingPointId", "responsibleName" ],
        "type" : "object",
        "properties" : {
          "visitId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "clientOperationId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "operatingPointId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "responsibleName" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "checkinAt" : {
            "type" : "string"
          },
          "checkinLatitude" : {
            "format" : "double",
            "type" : "number"
          },
          "checkinLongitude" : {
            "format" : "double",
            "type" : "number"
          }
        }
      },
...
      "UUID" : {
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
        "type" : "string"
      },
...
      "VisitResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "operatingPointId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "operatingPointName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "responsibleName" : {
            "type" : "string"
          },
          "checkinAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "checkoutAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "cashCollectedCents" : {
            "format" : "int64",
            "type" : "integer"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "description" : "Authentication",
        "openIdConnectUrl" : "http://localhost:8180/realms/gruahub/.well-known/openid-configuration"
      }
    }
  }
}